basic_numpy/project/requirements.txt
PyYAML
typer
numpy